Fix error #124
When creating a new Event with badly formatted images, a new event was
created without the images but the user was not notified of this. This
led to an event being created multiple times. With this commit the user
is redirected to an edit event view where he can change the images.